Implement `AlertBlockSyntax` using `BlockquoteSyntax`
As I understand the alert-block syntax, it is an extension on the
orignal blockquote support in markdown. Basically, it's a blockquote
that is rendered differently.

That means that if alert-block syntax isn't supported a markdown parser
will render it as a blockquote. This provides a nice graceful fallback.
It also means that if you're implementing it, you really should
implement it the same way you implement blockquote parsing.
